requirements.txt
numpy==1.19.5
sentence-transformers==2.2.2
scikit-image==0.17.2
scikit-learn==0.24.2
tqdm>=4.59
torchmetrics==0.8.2
tensorboard==2.4.0
hydra-core
seaborn
matplotlib
pydensecrf
wget
pytorch_lightning